requirements.txt
django==4.1
django-taggit==3.0.0
markdown==3.4.1
psycopg2-binary==2.9.3